Make the minimum user password length configurable
Previously, all passwords had to be at least 12 characters long.

Some users use a slightly shorter password, so some instances could wish
to reduce the password length limit to a slightly lesser value.

This makes the minimum password length configurable in the .env file,
with the MIN_PASSWORD_LENGTH variable, that defaults to 12 to keep the
previous behaviour by default.