fix broken accessibility for search field
This PR fixes the incorrect use of `role=search` on the `<input>` element.  The `search` role is not a type of text input, and its use here was negating the semantics of the text field's ability to announce itself as a text field to screen readers.  The proper place for a `role=search` is on the `<form>` element that contains this input, so I have placed it there in this fix.

There are likely other accessibility improvements that could be made in this file, but this was an easy fix that I figured I could make a quick PR for.  

Thanks