chore: adds node version
Adds a .node-version file to the repo which allows various developer tools
such as fnm and others to load the correct version of Node

I am inferring that Node v14 is the version being used.
When running `npm install` with Node >=16, npm states the
lockfile was run with an older version of npm and needs to
apply a one off upgrade. The lockfile version generated by
npm changed from v1 => v2 with npm v7.
https://github.blog/2021-02-02-npm-7-is-now-generally-available/#changes-to-the-lockfile

A separate and future work item would be to bump the
project to use the latest version of Node.