Add support for OIDC login and authentication.
If `OIDC_ENABLED=true` is set in the environment, the usual login
and registration UI will not be available and only the SSO link is
visible.

If the OIDC provider says the user is in group `admin`, they will be an
admin on pixelfed.

The OpenID Connect library comes from Cabinet Office Digital Development
and is included in the tree.  It might be possible to fetch it via
the composer package, although I'm not sure how to do that.