docker: Remove bootstrap volume and streamline startup
This patch modifies the `start.apache.sh` script to perform the
initial setup on first boot of the container so that the database
and all of the various keys are applied.

The patch also moves the storage, db, and redis volumes into a local
`data` directory for easier management and removes the bootstrap volume
since it did not seem necessary for normal operation.