Change the collation for the hashtags table
Changes the collation for the hashtags table to utf8mb4_unicode_520_ci
when the driver is MySQL or MariaDB. Otherwise the default collation of
utf8mb4_unicode_ci is used which conflates all characters outside of the
basic multilingual plane. This meant that any hashtag that is using a
script not in the BMP ends up matching with any other hashtag with a
script outside the BMP if it has the same length.